Abstract out calendar component
